polymarket_rpi/requirements.txt
aiohttp>=3.8.0
pyzmq>=25.0.0
